Value Verification.!

+++ main form.py
@@ -62,12 +62,15 @@ btn_per_select_all = tk.Button(master=form, text='همه کاربر ها', comma
 
 
 def select_person():
-    # جستجوی یک رکورد در پایگاه با استفاده از مقدار های ارسال شده
-    record = select_record(tbl_name='tbl_persons', id=txb_per_id.get())
-    # خالی کردن تکست باکس
-    txb_per_name.delete(0, tk.END)
-    # افزودن رکورد دریافت شده به تکست باکس
-    txb_per_name.insert(tk.END, record[1])
+    if not txb_per_id.get().isdigit():
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه باید عدد باشد')
+    else:
+        # جستجوی یک رکورد در پایگاه با استفاده از مقدار های ارسال شده
+        record = select_record(tbl_name='tbl_persons', id=txb_per_id.get())
+        # خالی کردن تکست باکس
+        txb_per_name.delete(0, tk.END)
+        # افزودن رکورد دریافت شده به تکست باکس
+        txb_per_name.insert(tk.END, record[1])
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_per_select = tk.Button(master=form, text='جستجو', command=select_person)
 
@@ -87,29 +90,35 @@ btn_per_insert = tk.Button(master=form, text='افزودن کاربر', command=
 
 
 def update_person():
-    # ویرایش یک رکورد در پایگاه داده
-    # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
-    query = update_record(tbl_name='tbl_persons', target_id=txb_per_id.get(), p_name=txb_per_name.get())
-    if query:
-        # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
-        select_all_persons()
-    elif query[0] is False:
-        # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
-        messagebox.showerror('کاربر ویرایش نشد', query[1])
+    if not txb_per_id.get().isdigit():
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه باید عدد باشد')
+    else:
+        # ویرایش یک رکورد در پایگاه داده
+        # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
+        query = update_record(tbl_name='tbl_persons', target_id=txb_per_id.get(), p_name=txb_per_name.get())
+        if query:
+            # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
+            select_all_persons()
+        elif query[0] is False:
+            # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
+            messagebox.showerror('کاربر ویرایش نشد', query[1])
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_per_update = tk.Button(master=form, text='ویرایش کاربر', command=update_person)
 
 
 def delete_person():
-    # حذف یک رکورد از پایگاه داده
-    # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
-    query = delete_record(tbl_name='tbl_persons', target_id=txb_per_id.get())
-    if query:
-        # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
-        select_all_persons()
-    elif query[0] is False:
-        # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
-        messagebox.showerror('کاربر حذف نشد', query[1])
+    if not txb_per_id.get().isdigit():
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه باید عدد باشد')
+    else:
+        # حذف یک رکورد از پایگاه داده
+        # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
+        query = delete_record(tbl_name='tbl_persons', target_id=txb_per_id.get())
+        if query:
+            # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
+            select_all_persons()
+        elif query[0] is False:
+            # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
+            messagebox.showerror('کاربر حذف نشد', query[1])
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_per_delete = tk.Button(master=form, text='حذف کاربر', command=delete_person)
 
@@ -126,6 +135,8 @@ def select_all_vehicles():
         ایدی کاربر مالک خودرو را جداسازی کرده کاربر را یافته و درون متغیر جای داده
         """
         id = int(record[2])
+        # if not isinstance(id, int) and id[0] is False:
+        #     delete_record(tbl_name='tbl_vehicles', target_id=record[0])
         # جستجوی کاربر صاحب خودرو در پایگاه داده و قرار دادن در متغیر
         person = select_record(tbl_name='tbl_persons', id=id)
         # افزودن آیتم به لیست باکس(پارامتر اول شمارنده حلقه و دوم یک رشته حاوی رکورد های دریافت شده از پایگاه داده)
@@ -135,58 +146,70 @@ btn_car_select_all = tk.Button(master=form, text='همه خودرو ها', comma
 
 
 def select_vehicle():
-    # جستجوی یک رکورد در پایگاه با استفاده از مقدار های ارسال شده
-    record = select_record(tbl_name='tbl_vehicles', id=txb_car_id.get())
-    # خالی کردن تکست باکس
-    txb_car_name.delete(0, tk.END)
-    # افزودن رکورد دریافت شده به تکست باکس
-    txb_car_name.insert(tk.END, record[1])
-    # خالی کردن تکست باکس
-    txb_car_person__id.delete(0, tk.END)
-    # افزودن رکورد دریافت شده به تکست باکس
-    txb_car_person__id.insert(tk.END, record[2])
+    if not txb_car_id.get().isdigit():
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه باید عدد باشد')
+    else:
+        # جستجوی یک رکورد در پایگاه با استفاده از مقدار های ارسال شده
+        record = select_record(tbl_name='tbl_vehicles', id=txb_car_id.get())
+        # خالی کردن تکست باکس
+        txb_car_name.delete(0, tk.END)
+        # افزودن رکورد دریافت شده به تکست باکس
+        txb_car_name.insert(tk.END, record[1])
+        # خالی کردن تکست باکس
+        txb_car_person__id.delete(0, tk.END)
+        # افزودن رکورد دریافت شده به تکست باکس
+        txb_car_person__id.insert(tk.END, record[2])
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_car_select = tk.Button(master=form, text='جستجو خودرو', command=select_vehicle)
 
 def insert_vehicle():
-    # افزودن یک رکورد به پایگاه داده
-    # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
-    query = insert_record(tbl_name='tbl_vehicles', v_name=txb_car_name.get(), v_pid=txb_car_person__id.get())
-    if query:
-        # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
-        select_all_vehicles()
-    elif query[0] is False:
-        # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
-        messagebox.showwarning('خودرو افزوده نشد', query[1])
+    if txb_car_person__id.get().isdigit():
+        # افزودن یک رکورد به پایگاه داده
+        # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
+        query = insert_record(tbl_name='tbl_vehicles', v_name=txb_car_name.get(), v_pid=txb_car_person__id.get())
+        if query:
+            # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
+            select_all_vehicles()
+        elif query[0] is False:
+            # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
+            messagebox.showwarning('خودرو افزوده نشد', query[1])
+    else:
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه ها باید عدد باشد')
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_car_insert = tk.Button(master=form, text='افزودن خودرو', command=insert_vehicle)
 
 
 def update_vehicle():
-    # ویرایش یک رکورد در پایگاه داده
-    # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
-    query = update_record(tbl_name='tbl_vehicles', target_id=txb_car_id.get(), 
-                        v_name=txb_car_name.get(), v_pid=txb_car_person__id.get())
-    if query:
-        # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
-        select_all_vehicles()
-    elif query[0] is False:
-        # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
-        messagebox.showerror('خودرو ویرایش نشد', query[1])
+    if txb_car_id.get().isdigit() and txb_car_person__id.get().isdigit():
+        # ویرایش یک رکورد در پایگاه داده
+        # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
+        query = update_record(tbl_name='tbl_vehicles', target_id=txb_car_id.get(), 
+                            v_name=txb_car_name.get(), v_pid=txb_car_person__id.get())
+        if query:
+            # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
+            select_all_vehicles()
+        elif query[0] is False:
+            # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
+            messagebox.showerror('خودرو ویرایش نشد', query[1])
+    else:
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه ها باید عدد باشد')
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_car_update = tk.Button(master=form, text='ویرایش خودرو', command=update_vehicle)
 
 
 def delete_vehicle():
-    # حذف یک رکورد از پایگاه داده
-    # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
-    query = delete_record(tbl_name='tbl_vehicles', target_id=txb_car_id.get())
-    if query:
-        # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
-        select_all_vehicles()
-    elif query[0] is False:
-        # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
-        messagebox.showerror('خودرو حذف نشد', query[1])
+    if not txb_car_id.get().isdigit():
+        messagebox.showerror('مقدار اشتباه', 'مقدار شناسه باید عدد باشد')
+    else:
+        # حذف یک رکورد از پایگاه داده
+        # تابع به برای ما یک مقدار درست یا غلط و متن ارور را باز میگرداند
+        query = delete_record(tbl_name='tbl_vehicles', target_id=txb_car_id.get())
+        if query:
+            # اگر مقدار بازگردانده شده درست بود تابع نمایش همه ی کاربر ها اجرا شود
+            select_all_vehicles()
+        elif query[0] is False:
+            # اگر مقدار بازگردانده شده غلط بود ارور در مسیج باکس نمایش داده شود
+            messagebox.showerror('خودرو حذف نشد', query[1])
 # نمونه سازی از یک دکمه و متصل کردن آن به تابع بالا
 btn_car_delete = tk.Button(master=form, text='حذف خودرو', command=delete_vehicle)
